Revert "[MBVM-101] Upgrade MB Solr version to 4.0.0 (#304)"
This reverts commit 70aaf8cf57ee59bb39e19cb936cbea3400046ebe,
thus allowing mirror owner to keep running legacy versions of
Solr 7 / Python 2 / Docker Compose 1 for a transition period.

+++ build/musicbrainz/scripts/fetch-dump.sh
@@ -13,12 +13,13 @@ SCRIPT_NAME=$(basename "$0")
 HELP=$(cat <<EOH
 Usage: $SCRIPT_NAME [<options>] <target>
 
-Fetch dump files of the MusicBrainz Postgres database.
+Fetch dump files of the MusicBrainz database and/or search indexes.
 
 Targets:
-  indexed       Fetch latest database's replicated tables with Solr backups available.
+  both          Fetch latest search dump with replica dump of the same day.
   replica       Fetch latest database's replicated tables only.
   sample        Fetch latest database's sample only.
+  search        Fetch latest search indexes only.
 
 Options:
   --base-download-url <url>     Specify URL of a MetaBrainz/MusicBrainz download server.
@@ -39,7 +40,7 @@ EOH
 while [[ $# -gt 0 ]]
 do
 	case "$1" in
-		indexed | replica | sample )
+		both | replica | sample | search )
 			if [[ -n $TARGET ]]
 			then
 				echo >&2 "$SCRIPT_NAME: only one target argument can be given"
@@ -101,9 +102,7 @@ fi
 
 if [[ $TARGET != sample &&
 	! -a "$DB_DUMP_DIR/.for-commercial-use" &&
-	! -a  "$DB_DUMP_DIR/.for-non-commercial-use" &&
-	! -a "$SEARCH_DUMP_DIR/.for-commercial-use" &&
-	! -a  "$SEARCH_DUMP_DIR/.for-non-commercial-use"
+	! -a  "$DB_DUMP_DIR/.for-non-commercial-use"
 ]]
 then
 	prompt=$(cat <<-EOQ
@@ -156,24 +155,62 @@ fi
 
 BASE_DOWNLOAD_URL="${BASE_FTP_URL:-$BASE_DOWNLOAD_URL}"
 
+# Fetch latest search indexes
+
+if [[ $TARGET =~ ^(both|search)$ ]]
+then
+	echo "$(date): Fetching search indexes dump..."
+	PREVIOUS_DUMP_TIMESTAMP=''
+	if [[ -a "$SEARCH_DUMP_DIR/LATEST" ]]
+	then
+		PREVIOUS_DUMP_TIMESTAMP=$(<"$SEARCH_DUMP_DIR/LATEST")
+		rm -f "$SEARCH_DUMP_DIR/LATEST"
+	fi
+	"${WGET_CMD[@]}" -nd -nH -P "$SEARCH_DUMP_DIR" \
+		"${BASE_DOWNLOAD_URL}/data/search-indexes/LATEST"
+	DUMP_TIMESTAMP=$(<"$SEARCH_DUMP_DIR/LATEST")
+	if [[ $PREVIOUS_DUMP_TIMESTAMP != "$DUMP_TIMESTAMP" ]]
+	then
+		find "$SEARCH_DUMP_DIR" \
+			! -path "$SEARCH_DUMP_DIR" \
+			! -path "$DB_DUMP_DIR/.for-commercial-use" \
+			! -path "$DB_DUMP_DIR/.for-non-commercial-use" \
+			! -path "$SEARCH_DUMP_DIR/LATEST" \
+			-delete
+	fi
+	"${WGET_CMD[@]}" -nd -nH -c -r -P "$SEARCH_DUMP_DIR" \
+		--accept 'MD5SUMS,*.tar.zst' --no-parent --relative \
+		"${BASE_DOWNLOAD_URL}/data/search-indexes/$DUMP_TIMESTAMP/"
+	echo "$(date): Checking MD5 sums..."
+	cd "$SEARCH_DUMP_DIR" && md5sum -c MD5SUMS && cd - >/dev/null
+	if [[ $TARGET == search ]]
+	then
+		echo 'Done fetching search indexes dump'
+		exit 0 # EX_OK
+	fi
+fi
+
 # Prepare to fetch database dump
 
-echo "$(date): Fetching database dump..."
+if [[ $TARGET != search ]]
+then
+	echo "$(date): Fetching database dump..."
 
-PREVIOUS_DUMP_TIMESTAMP=''
-if [[ -a "$DB_DUMP_DIR/LATEST" ]]
-then
-	PREVIOUS_DUMP_TIMESTAMP=$(<"$DB_DUMP_DIR/LATEST")
-	rm -f "$DB_DUMP_DIR/LATEST"
-fi
-if [[ -a "$DB_DUMP_DIR/LATEST-WITH-SEARCH-INDEXES" ]]
-then
-	PREVIOUS_DUMP_TIMESTAMP=$(<"$DB_DUMP_DIR/LATEST-WITH-SEARCH-INDEXES")
-	rm -f "$DB_DUMP_DIR/LATEST-WITH-SEARCH-INDEXES"
+	PREVIOUS_DUMP_TIMESTAMP=''
+	if [[ -a "$DB_DUMP_DIR/LATEST" ]]
+	then
+		PREVIOUS_DUMP_TIMESTAMP=$(<"$DB_DUMP_DIR/LATEST")
+		rm -f "$DB_DUMP_DIR/LATEST"
+	fi
+	if [[ -a "$DB_DUMP_DIR/LATEST-WITH-SEARCH-INDEXES" ]]
+	then
+		PREVIOUS_DUMP_TIMESTAMP=$(<"$DB_DUMP_DIR/LATEST-WITH-SEARCH-INDEXES")
+		rm -f "$DB_DUMP_DIR/LATEST-WITH-SEARCH-INDEXES"
+	fi
 fi
 
 case "$TARGET" in
-	indexed | replica )
+	both | replica )
 		DB_DUMP_REMOTE_DIR=data/fullexport
 		DB_DUMP_FILES=(
 			mbdump.tar.bz2
@@ -192,15 +229,11 @@ case "$TARGET" in
 		;;
 esac
 
-if [[ $TARGET == indexed ]]
+if [[ $TARGET == both ]]
 then
 	# Find latest database dump corresponding to search indexes
 
-	SEARCH_DUMP_TIMESTAMP=$(
-		"${WGET_CMD[@]}" -q -O - \
-			"${BASE_DOWNLOAD_URL}/data/solr-backups/LATEST"
-	)
-	SEARCH_DUMP_DAY="${SEARCH_DUMP_TIMESTAMP/-*}"
+	SEARCH_DUMP_DAY="${DUMP_TIMESTAMP/-*}"
 	rm -f "$DB_DUMP_DIR/index.html"
 	"${WGET_CMD[@]}" --force-html -O "$DB_DUMP_DIR/index.html" -P "$DB_DUMP_DIR" \
 		"${BASE_DOWNLOAD_URL}/$DB_DUMP_REMOTE_DIR/"
@@ -211,7 +244,8 @@ then
 	)
 	rm -f "$DB_DUMP_DIR/index.html"
 	echo "$DUMP_TIMESTAMP" >> "$DB_DUMP_DIR/LATEST-WITH-SEARCH-INDEXES"
-else
+elif [[ $TARGET != search ]]
+then
 	# Just find latest database dump
 
 	"${WGET_CMD[@]}" -nd -nH -P "$DB_DUMP_DIR" \
@@ -221,20 +255,23 @@ fi
 
 # Remove previously downloaded files if obsolete
 
-if [[ $PREVIOUS_DUMP_TIMESTAMP != "$DUMP_TIMESTAMP" ]]
+if [[ $TARGET != search ]]
 then
-	find "$DB_DUMP_DIR" \
-		! -path "$DB_DUMP_DIR" \
-		! -path "$DB_DUMP_DIR/.for-commercial-use" \
-		! -path "$DB_DUMP_DIR/.for-non-commercial-use" \
-		! -path "$DB_DUMP_DIR/LATEST" \
-		! -path "$DB_DUMP_DIR/LATEST-WITH-SEARCH-INDEXES" \
-		-delete
+	if [[ $PREVIOUS_DUMP_TIMESTAMP != "$DUMP_TIMESTAMP" ]]
+	then
+		find "$DB_DUMP_DIR" \
+			! -path "$DB_DUMP_DIR" \
+			! -path "$DB_DUMP_DIR/.for-commercial-use" \
+			! -path "$DB_DUMP_DIR/.for-non-commercial-use" \
+			! -path "$DB_DUMP_DIR/LATEST" \
+			! -path "$DB_DUMP_DIR/LATEST-WITH-SEARCH-INDEXES" \
+			-delete
+	fi
 fi
 
 # Actually fetch database dump
 
-if [[ $TARGET =~ ^(indexed|replica)$ ]]
+if [[ $TARGET =~ ^(both|replica)$ ]]
 then
 	for F in MD5SUMS "${DB_DUMP_FILES[@]}"
 	do
